refactor(注册中心模块): 调整类的位置
Closes 更新结束
